Fix kernels

+++ cuda/rwkv7_clampw_v3_for_h100_alt.cu
@@ -9,7 +9,7 @@
     #include <cuda_bf16.h>
     using bf = __nv_bfloat16;
     #define to_float(u) (__bfloat162float(u))
-    #define to_bf(u) (__float2bfloat16_rn(u))
+    #define to_bf(u) (__float2bfloat16(u))
 #endif
 
 using i64 = long long int;
